+++ resources/js/views/Measurements.vue
@@ -1,88 +1,65 @@
 <template>
-    <v-card>
+  <v-card>
+    <v-card-title>
+      Measurements
+      <v-spacer></v-spacer>
+      <v-text-field
+        v-model="search"
+        append-icon="mdi-magnify"
+        label="Search"
+        single-line
+        hide-details
+      ></v-text-field>
+    </v-card-title>
+    <v-data-table
+      :headers="headers"
+      :items="measurements"
+      :search="search"
+      :loading="loading"
+      loading-text="Loading Measurements... Please wait"
+    >
+      <template v-slot:item.action="{ item }">
+        <v-icon small class="mr-2" @click="editItem(item)">mdi-pencil</v-icon>
+        <v-icon small @click="deleteItem(item)">mdi-delete</v-icon>
+      </template>
+      <template v-slot:no-data>
+        <v-btn color="primary" @click="initialize">Reset</v-btn>
+      </template>
+    </v-data-table>
+    <v-dialog v-model="dialog" max-width="500px">
+      <template v-slot:activator="{ on }">
+        <v-btn color="primary" dark class="mb-2" v-on="on">New Item</v-btn>
+      </template>
+      <v-card>
         <v-card-title>
-            Measurements
-            <v-spacer></v-spacer>
-            <v-text-field
-                    v-model="search"
-                    append-icon="mdi-magnify"
-                    label="Search"
-                    single-line
-                    hide-details
-            ></v-text-field>
+          <span class="headline">{{ formTitle }}</span>
         </v-card-title>
-        <v-data-table
-                :headers="headers"
-                :items="measurements"
-                :search="search"
-                :loading="loading"
-                loading-text="Loading Measurements... Please wait"
-        >
-            <template v-slot:item.action="{ item }">
-                <v-icon
-                        small
-                        class="mr-2"
-                        @click="editItem(item)"
-                >
-                    mdi-pencil
-                </v-icon>
-                <v-icon
-                        small
-                        @click="deleteItem(item)"
-                >
-                    mdi-delete
-                </v-icon>
-            </template>
-            <template v-slot:no-data>
-                <v-btn color="primary" @click="initialize">Reset</v-btn>
-            </template>
-        </v-data-table>
-        <v-dialog v-model="dialog" max-width="500px">
-            <template v-slot:activator="{ on }">
-                <v-btn color="primary" dark class="mb-2" v-on="on">New Item</v-btn>
-            </template>
-            <v-card>
-                <v-card-title>
-                    <span class="headline">{{ formTitle }}</span>
-                </v-card-title>
 
-                <v-card-text>
-                    <v-container>
-                        <v-row>
-                            <v-col cols="12" md="6">
-                                <v-text-field v-model="editedItem.name" label="Name"></v-text-field>
-                            </v-col>
-                            <v-col cols="12" md="6">
-                                <v-text-field v-model="editedItem.abbreviation" label="Abbreviation"></v-text-field>
-                            </v-col>
-                            <v-col cols="12" md="6">
-                                <v-select v-model="editedItem.type" :items="types" label="Type" item-text="name" item-value="id" return-object></v-select>
-                            </v-col>
-                        </v-row>
-                    </v-container>
-                </v-card-text>
+        <v-card-text>
+          <v-container>
+            <v-row>
+              <v-col cols="12" md="6">
+                <v-text-field v-model="editedItem.name" label="Name"></v-text-field>
+              </v-col>
+              <v-col cols="12" md="6">
+                <v-text-field v-model="editedItem.abbreviation" label="Abbreviation"></v-text-field>
+              </v-col>
+            </v-row>
+          </v-container>
+        </v-card-text>
 
-                <v-card-actions>
-                    <v-spacer></v-spacer>
-                    <v-btn color="blue darken-1" text @click="close">Cancel</v-btn>
-                    <v-btn color="blue darken-1" text @click="save">Save</v-btn>
-                </v-card-actions>
-            </v-card>
-        </v-dialog>
-        <v-snackbar
-                v-model="snackbar"
-                :timeout="snackbarTimeout"
-        >
-            {{ snackbarText }}
-            <v-btn
-                    color="blue"
-                    text
-                    @click="snackbar = false"
-            >
-                Close
-            </v-btn>
-        </v-snackbar>
-    </v-card>
+        <v-card-actions>
+          <v-spacer></v-spacer>
+          <v-btn color="blue darken-1" text @click="close">Cancel</v-btn>
+          <v-btn color="blue darken-1" text @click="save">Save</v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+    <v-snackbar v-model="snackbar" :timeout="snackbarTimeout">
+      {{ snackbarText }}
+      <v-btn color="blue" text @click="snackbar = false">Close</v-btn>
+    </v-snackbar>
+  </v-card>
 </template>
 
 <script>
@@ -92,11 +69,12 @@ export default {
       search: "",
       headers: [
         {
-          text: "Name",
+          text: "id",
           align: "left",
           sortable: false,
-          value: "name"
+          value: "id"
         },
+        { text: "Name", value: "name" },
         { text: "Abbreviation", value: "abbreviation" },
         { text: "Type", value: "type.name" },
         { text: "Created at", value: "created_at" },
@@ -108,8 +86,7 @@ export default {
       editedIndex: -1,
       editedItem: {
         name: "",
-        abbreviation: "",
-        type: ""
+        abbreviation: ""
       },
       defaultItem: {
         name: ""
@@ -165,11 +142,13 @@ export default {
 
     deleteItem(item) {
       const index = this.measurements.indexOf(item);
-      confirm("Are you sure you want to delete this measurement") &&
-        this.measurements.splice(index, 1);
-      axios.delete("api/measurements/" + item.id);
-      this.snackbarText = "Measurement deleted";
-      this.snackbar = true;
+      if (confirm("Are you sure you want to delete this measurement")) {
+        if (this.measurements.splice(index, 1)) {
+          axios.delete("api/measurements/" + item.id);
+          this.snackbarText = "Measurement deleted";
+          this.snackbar = true;
+        }
+      }
     },
 
     close() {
